refactor(chats): split useChatsStore into slice modules under src/stores/chats
Extract types, shared helpers, AI/auth/rooms slices, and persist wiring into
dedicated files; useChatsStore.ts re-exports. Wire forceLogout via store ref
for restoreSessionFromCookie. Extend guard wiring test to scan all chat store sources.
